+++ footer.php
@@ -1,23 +1,27 @@
+<?php
+/**
+ * Theme footer template.
+ *
+ * @package TailPress
+ */
+?>
+        </main>
 
-</main>
+        <?php do_action('tailpress_content_end'); ?>
+    </div>
 
-<?php do_action( 'pekebyte_one_content_end' ); ?>
+    <?php do_action('tailpress_content_after'); ?>
 
-</div>
-
-<?php do_action( 'pekebyte_one_content_after' ); ?>
-
-<footer id="colophon" class="site-footer py-12" role="contentinfo">
-	<?php do_action( 'pekebyte_one_footer' ); ?>
-
-	<div class="container mx-auto text-center">
-		&copy; <?php echo date_i18n( 'Y' );?> - <?php echo get_bloginfo( 'name' );?>
-	</div>
-</footer>
-
+    <footer id="colophon" class="bg-light/50 mt-12" role="contentinfo">
+        <div class="container mx-auto py-12">
+            <?php do_action('tailpress_footer'); ?>
+            <div class="text-sm text-zinc-700">
+                &copy; <?php echo esc_html(date_i18n('Y')); ?> - <?php bloginfo('name'); ?>
+            </div>
+        </div>
+    </footer>
 </div>
 
 <?php wp_footer(); ?>
-
 </body>
 </html>