+++ resources/js/app.js
@@ -1,8 +1,11 @@
-// Navigation toggle
 window.addEventListener('load', function () {
-      let main_navigation = document.querySelector('#primary-menu');
-      document.querySelector('#primary-menu-toggle').addEventListener('click', function (e) {
-            e.preventDefault();
-            main_navigation.classList.toggle('hidden');
-      });
-});
+    let mainNavigation = document.getElementById('primary-navigation')
+    let mainNavigationToggle = document.getElementById('primary-menu-toggle')
+
+    if(mainNavigation && mainNavigationToggle) {
+        mainNavigationToggle.addEventListener('click', function (e) {
+            e.preventDefault()
+            mainNavigation.classList.toggle('hidden')
+        })
+    }
+})