+++ single.php
@@ -1,21 +1,24 @@
-<?php get_header(); ?>
+<?php
+/**
+ * Single post template file.
+ *
+ * @package TailPress
+ */
 
-	<div class="container my-8 mx-auto">
+get_header();
+?>
 
-	<?php if ( have_posts() ) : ?>
+<div class="container my-8 mx-auto">
+    <?php if (have_posts()): ?>
+        <?php while (have_posts()): the_post(); ?>
+            <?php get_template_part('template-parts/content', 'single'); ?>
 
-		<?php
-		while ( have_posts() ) :
-			the_post();
-			?>
-
-			<?php get_template_part( 'template-parts/content', 'single' ); ?>
-
-		<?php endwhile; ?>
-
-	<?php endif; ?>
-
-	</div>
+            <?php if (comments_open() || get_comments_number()): ?>
+                <?php comments_template(); ?>
+            <?php endif; ?>
+        <?php endwhile; ?>
+    <?php endif; ?>
+</div>
 
 <?php
 get_footer();