+++ template-parts/content-single.php
@@ -1,36 +1,21 @@
 <article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
+    <header class="mx-auto flex max-w-5xl flex-col text-center">
+        <h1 class="mt-6 text-5xl font-medium tracking-tight [text-wrap:balance] text-zinc-950 sm:text-6xl"><?php the_title(); ?></h1>
 
-	<header class="entry-header mb-4 flex flex-wrap justify-between p-5">
-		<?php
-			/** In order to display the correct Icon will need the last category */
-			$last_child_category = get_the_last_child_category(get_the_ID());
-		?>
-		<div class="w-1/12 self-center"><img src="<?php echo get_cat_icon($last_child_category) ?>" alt="<?php printf( __( 'Image of category %s', 'pekebyte-one' ), $last_child_category->name ); ?>" /></div>
-		<div class="w-11/12 px-5">
-			<?php the_title( sprintf( '<h1 class="entry-title text-2xl lg:text-5xl font-extrabold leading-tight mb-1"><a href="%s" rel="bookmark">', esc_url( get_permalink() ) ), '</a></h1>' ); ?>
-			<div class="w-full flex flex-wrap">
-				<time datetime="<?php echo get_the_date( 'c' ); ?>" itemprop="datePublished" class="text-sm text-darker-light-blue"><?php echo get_the_date(); ?></time>
-				<div class="w-full flex sm:w-auto"><div class="ml-2 hidden sm:block text-sm text-darker-light-blue">-</div>
-				<div class="sm:mt-0 sm:ml-2 text-sm text-darker-light-blue"><?php get_the_reading_time(get_the_ID()); ?></div></div>
-			</div>
-		</div>
-	</header>
+        <?php if(! is_page()): ?>
+            <time datetime="<?php echo get_the_date( 'c' ); ?>" itemprop="datePublished" class="order-first text-sm text-zinc-950"><?php echo get_the_date(); ?></time>
 
-	<div class="entry-content max-w-[960px] mx-auto">
-		<?php the_content(); ?>
+            <p class="mt-6 text-sm font-semibold text-zinc-950">by <?php the_author(); ?></p>
+        <?php endif; ?>
+    </header>
 
-		<?php
-			wp_link_pages(
-				array(
-					'before'      => '<div class="page-links"><span class="page-links-title">' . __( 'Pages:', 'pekebyte-one' ) . '</span>',
-					'after'       => '</div>',
-					'link_before' => '<span>',
-					'link_after'  => '</span>',
-					'pagelink'    => '<span class="screen-reader-text">' . __( 'Page', 'pekebyte-one' ) . ' </span>%',
-					'separator'   => '<span class="screen-reader-text">, </span>',
-				)
-			);
-		?>
-	</div>
+    <?php if(has_post_thumbnail()): ?>
+        <div class="mt-10 sm:mt-20 mx-auto max-w-4xl rounded-4xl bg-light overflow-hidden">
+            <?php the_post_thumbnail('large', ['class' => 'aspect-16/10 w-full object-cover']); ?>
+        </div>
+    <?php endif; ?>
 
+    <div class="entry-content mx-auto max-w-3xl mt-10 sm:mt-20">
+        <?php the_content(); ?>
+    </div>
 </article>